Bugfix for last commit.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5149 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -772,7 +772,7 @@ struct s_ecmWhitelist {
 };
 
 struct s_ecmWhitelistIdent {
-	uint16_t ident;
+	uint32_t ident;
 	struct s_ecmWhitelistLen *lengths;
 	struct s_ecmWhitelistIdent *next;
 };
